Web services (Java REST API) + PHP + HTML
Recipe book is the online web application made in Java REST API combined
with PHP anf HTML.

All the crud operation are performed in Java API and displayed in the
PHP page.

In this project API work as back end while PHP as missle tier and HTML,
CSS, Java script as front end.

This is my SPM project which i developed in my masters.